[FIXED] Animation system - critical bug
1. Critical bug fixed
2. Warnings resolved except for external warnings. (Just unreferenced parameter)

+++ engine/Graphics/AnimationSystem.h
@@ -229,8 +229,8 @@ namespace Framework {
 		Framework::EntityManager* entityManager;
         float max_static_threshold;
         float default_zero;
-        float anim_current_frame;
-        float anim_frame_mod;
+        int anim_current_frame;
+        int anim_frame_mod;
 	};
 
 } // namespace Framework
